screener_server/requirements.txt

Flask==1.1.2
Flask-Cors==3.0.10
numpy==1.19.2
psycopg2==2.7.7
python-dotenv==0.18.0
SQLAlchemy==1.4.2